Higher Contrast Option
I know you said you didn't want to have a bunch of settings for every interface option, but I thought I'd submit this just in case (the work is done, so no big deal for me). This time instead of a single small change I create two new color options to make the changes global. Basically the various g_70 settings got a value that's actually 80% and the various g_30 settings got a value that's actually 50%. If you don't want to do that, that's absolutely fine. This change makes it way easier for me to just drop one file in after a skin update to get back what I need to be able to see things better.