Slightly Larger Font
OK, new try at this.  I went through and find all the places with either multiline or textbox so I'd know the places where a larger font would cause problems.  I didn't change any of the lyric fonts (that handled all the multiline issues), and I didn't change plotfont (which handled a bunch of stuff). Where I did increase a font size that was in a textbox, I created a new font style ending in _abs and used that font style with a consistent size between the two font sets.  I found two "regular" text lines that were too crowded larger (the info labels under titles and the info under the seekbar) and also made those absolute.  Sharing mostly becasue the work is done now. If this still doesn't work for you, that's OK. It's an easy update if I have to maintain it on my own.