Updated Kodi action calls
For Kodi 19, the XBMC. format for action item calls has been removed (it was depreciated some time ago but continued working through Kodi 18). This change updates all the calls. The updated calls will work in Kodi 18 and are required for Kodi 19. One note: to get the power menu to update properly, I had to delete the powermenu entry from the script.skinshortcuts addondata directory so that the skin would regenerate a new one. I'm not sure if there's a better way to handle that.